Claims and Authentication

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -49,19 +50,20 @@ public class AuctionsController : ControllerBase
     return _mapper.Map<AuctionDTO>(auction);
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<AuctionDTO>> CreateAuction(CreateAuctionDTO auctionDTO)
   {
     var auction = _mapper.Map<Auction>(auctionDTO);
-    //TODO add current user as seller
-    auction.Seller = "test";
+    
+    auction.Seller = User.Identity.Name;
 
     _context.Auctions.Add(auction);
 
     var newAuction = _mapper.Map<AuctionDTO>(auction);
 
     await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
-    
+
     var result = await _context.SaveChangesAsync() > 0;
 
     if (!result) return BadRequest("Could not save changes to DB");
@@ -69,7 +71,7 @@ public class AuctionsController : ControllerBase
 
     return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
   }
-
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
   {
@@ -77,7 +79,7 @@ public class AuctionsController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    //TODO check seller == username
+    if(auction.Seller != User.Identity.Name) return Forbid("You are not the seller of this auction");
 
     auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;
     auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
@@ -94,6 +96,7 @@ public class AuctionsController : ControllerBase
     return BadRequest("Could not update auction");
   }
 
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<ActionResult> DeleteAuction(Guid id)
   {
@@ -101,7 +104,7 @@ public class AuctionsController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    //TODO check seller == username
+    if(auction.Seller != User.Identity.Name) return Forbid("You are not the seller of this auction");
 
     _context.Auctions.Remove(auction);
